+++ src/main/java/com/stevekung/replayfov/FovPositionKeyframe.java
@@ -4,7 +4,7 @@ import com.replaymod.replaystudio.pathing.change.Change;
 
 public interface FovPositionKeyframe
 {
-    void addPositionKeyframe(long time, double posX, double posY, double posZ, float yaw, float pitch, float roll, float fov, int spectated);
+    void addPositionKeyframe(long time, double posX, double posY, double posZ, float yaw, float pitch, float roll, int fov, int spectated);
 
-    Change updatePositionKeyframe(long time, double posX, double posY, double posZ, float yaw, float pitch, float roll, float fov);
+    Change updatePositionKeyframe(long time, double posX, double posY, double posZ, float yaw, float pitch, float roll, int fov);
 }
